New ERROR message when Norb!=1 or Nspin!=2
The old error message was a little measliding, just stating the condition that leads to stop. Now is subsituted with the De Morgan equivalent to its negation, with a brief comment following. 

TO DO? ->[this constraint, which is indeed valid for the genuine (original) Kane Mele model, could be removed in order to study generalizations of the model, e.g. adding a staggered on-site energy, which actually is already possible in the driver and allowed in the input file...]