DATABASE UPDATED + MPI_Bcast-bug fixed.
With this update the "mpi_bcast(converged)" bug has been removed,
following the update of DMFT_TOOLS. Only a selected number of
drivers has been updated to comply with the new form of the
DMFT_TOOLS library.
In few lines:
* all DMFT_TOOLS related procedure DO NOT require MPI communicator
anymore, nor WTK k-point weight array, to be passed as input.
=> the input are identical for MPI and serial calls. All the MPI
magic is done internally and silently.

* but for user-defined procedures, there is no more need to pre-pend
an if(master) condition to all DMFT_TOOLS procedures.

See the updated drivers for more informations.

About the "MPI_BCAST(CONVERGED)" BUG:
The bug was triggered by the concomitant existence of these three
conditions:
1. Optimized compilation [-O2]
2. convergence test performed by the master only [if(master)convergence=check_convergence..]
3. convergence info/exit conditions broadcasted to the nodes using Scifor wrapper:
[ Bcast_MPI(comm,converved)].
Presumably in a highly optimized compilation the bcasting of the logical variable
converged from the master to the nodes was anticipated by the do-while loop restart.
This caused the master (who knew the variable) to exit the loop, while the
remaining nodes to restart and hang forever. The master entering any procedure placed
after convergence produced either error or hang without exit.

Thanks to Francesca Paoletti, Gabriele Bellomia, Severino Adler for testing and
reporting about this bug.