Added driver for calculations on the Haldane model
This driver is supposed to be a running example on
performing DMFT calculations on the Haldane Model.
It implements a two-side unit cell on a honeycomb
lattice, where each side hosts one orbital and
explicitly uses the spin degree of freedom.
Therefore, AFM phases are accessible with this
driver.
It implements two versions:
1.) The real haldane model, where both spin
    channels are copies.
2.) A time-reversal symmetric version, where the
    non-interacting hamiltonian is time-reversal
    symmetric.